requirements.txt
apache-airflow==1.10.2
tenacity==4.12.0
python-dotenv==0.10.1
equests==2.21.0
